requirements.txt
aiohttp>=3.9.0
aiohttp-cors>=0.7.0
websockets>=12.0
bcrypt>=4.1.0
pyjwt>=2.8.0
aiofiles>=23.2.0
